refactor: extract task claims management into taskClaims.ts
- Extract all claim-related interfaces, helpers, and exported functions
  from taskFile.ts (1,702 lines) into new taskClaims.ts (544 lines)
- taskFile.ts re-exports via `export * from './taskClaims'` so all
  existing import paths continue to work unchanged
- taskFile.ts reduced from 1,702 to 1,158 lines
- Extracted: acquireClaim, releaseClaim, inspectClaimOwnership,
  inspectTaskClaimGraph, resolveStaleClaim plus 18 private helpers
